Move BotAccount out of MultiAccountManager.
