+++ PoGo.NecroBot.Logic/Event/BotSwitchedEvent.cs
@@ -1,10 +1,12 @@
-﻿namespace PoGo.NecroBot.Logic.Event
+﻿using PoGo.NecroBot.Logic.Model;
+
+namespace PoGo.NecroBot.Logic.Event
 {
     public class BotSwitchedEvent : IEvent
     {
-        private MultiAccountManager.BotAccount Account;
+        private BotAccount Account;
 
-        public BotSwitchedEvent(MultiAccountManager.BotAccount nextBot)
+        public BotSwitchedEvent(BotAccount nextBot)
         {
             Account = nextBot;
         }
