+++ PoGo.NecroBot.Logic/Exceptions/ActiveSwitchAccountManualException.cs
@@ -1,12 +1,13 @@
-﻿using System;
+﻿using PoGo.NecroBot.Logic.Model;
+using System;
 
 namespace PoGo.NecroBot.Logic.Exceptions
 {
     public class ActiveSwitchAccountManualException : Exception
     {
-        public MultiAccountManager.BotAccount RequestedAccount;
+        public BotAccount RequestedAccount;
 
-        public ActiveSwitchAccountManualException(MultiAccountManager.BotAccount requestedAccount)
+        public ActiveSwitchAccountManualException(BotAccount requestedAccount)
         {
             RequestedAccount = requestedAccount;
         }
