+++ PoGo.NecroBot.Logic/Exceptions/ActiveSwitchByPokemonException.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using PoGo.NecroBot.Logic.Event;
 using POGOProtos.Enums;
+using PoGo.NecroBot.Logic.Model;
 
 namespace PoGo.NecroBot.Logic.Exceptions
 {
@@ -9,7 +10,7 @@ namespace PoGo.NecroBot.Logic.Exceptions
         public double LastLatitude { get; set; }
         public double LastLongitude { get; set; }
         public PokemonId LastEncounterPokemonId { get; set; }
-        public MultiAccountManager.BotAccount Bot { get; set; }
+        public BotAccount Bot { get; set; }
         public bool Snipe { get; set; }
         public EncounteredEvent EncounterData { get; set; }
     }
